separated into apps and added first iteration of models
