+++ apps/core/models.py
@@ -21,33 +21,3 @@ class Cohort(models.Model):
     end_date = models.DateField()
     members = models.ManyToManyField("UserProfile")
     organizers = models.ManyToManyField("UserProfile")
-
-
-# Project
-class Project(models.Model):
-    """
-    A team of people working on one project.
-    """
-    name = models.CharField(max_length=255)
-    description = models.TextField()
-    # demo_link = models.URLField(blank=True, null=True)
-    # image = models.ImageField()
-    technologies = models.TextField(default="To be determined, mentor will update")
-    cohort = models.ManyToManyField(Cohort)
-    city = models.ManyToManyField(City)
-    # is_approved = models.BooleanField(default=False)
-    members = models.ManyToManyField("UserProfile")
-    mentors = models.ManyToManyField("UserProfile")
-
-    def save(self, *args, **kwargs):
-        # ensure all mentors are members
-        return super(Project, self).save(*args, **kwargs)
-
-
-# UserProfile
-class UserProfile(models.Model):
-    name = models.CharField(max_length=255)
-    description = models.TextField()
-    tech_to_learn = models.TextField("Technologies I want to learn more about")
-    tech_i_know = models.TextField("Technologies I know")
-    # default value for description (that includes template)
